Ver 0.88.0.0 : Table & Cloud  Faster & Larger
This version introduces a new Table format that has superior speed and
capabilities. This may result is simulations running an order of
magnitude faster. Also cloud issues with larger simulations were
examined and documentation was updated to connect users to resources
that may help solve these. Also a GUI bug with along history was fixed.